requirements.txt
streamlit
scikit-learn
numpy
pyserial
joblib